@@ -30,44 +30,44 @@
   },
   "homepage": "https://github.com/mohanadkaleia/onenode#readme",
   "dependencies": {
-    "axios": "^0.21.1",
-    "body-parser": "^1.18.2",
-    "chokidar": "^3.0.1",
-    "cors": "^2.8.4",
+    "axios": "^1.6.7",
+    "body-parser": "^1.20.2",
+    "chokidar": "^3.5.3",
+    "cors": "^2.8.5",
     "delivery": "0.0.6",
-    "express": "^4.17.3",
-    "file-type": "^16.5.4",
-    "morgan": "^1.9.1",
-    "mysql": "^2.17.1",
-    "node-datetime": "^2.0.6",
+    "express": "^4.18.2",
+    "file-type": "^18.7.0",
+    "morgan": "^1.10.0",
+    "mysql": "^2.18.1",
+    "node-datetime": "^2.1.2",
     "path": "^0.12.7",
-    "public-ip": "^2.3.5",
-    "read-chunk": "^2.1.0",
-    "simple-peer": "^9.0.0",
-    "socket.io-client": "^2.1.1",
+    "public-ip": "^6.0.1",
+    "read-chunk": "^4.0.3",
+    "simple-peer": "^9.11.1",
+    "socket.io-client": "^4.7.4",
     "tree-model": "^1.0.7",
-    "vue": "^3.0.0",
-    "vue-router": "^3.0.1",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.3.12",
-    "websocket": "^1.0.24",
-    "winston": "^2.4.0",
-    "wrtc": "^0.3.2",
-    "ws": "^6.1.2",
-    "yargs": "^11.0.0"
+    "vue": "^3.4.15",
+    "vue-router": "^4.2.5",
+    "webpack": "^5.90.1",
+    "webpack-cli": "^5.1.4",
+    "websocket": "^1.0.34",
+    "winston": "^3.11.0",
+    "ws": "^8.16.0",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^2.1.1",
-    "file-loader": "^1.1.4",
-    "vue-loader": "^15.4.2",
-    "vue-template-compiler": "^2.4.4",
-    "webpack": "^4.8.3",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.7.1"
+    "@babel/cli": "^7.26.4",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
+    "@vue/compiler-sfc": "^3.4.15",
+    "babel-loader": "^9.2.1",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.10.0",
+    "file-loader": "^6.2.0",
+    "vue-loader": "^17.4.2",
+    "vue-style-loader": "^4.1.3",
+    "webpack": "^5.90.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
